feat: add more config entries to our localized default build rc json config file.
The srcDir specifies the project's main source code folder. It defaults to src. the
mainModuleFileName specifies the main entry file that would the library library's node.js main entry
file or export. the mainModuleName is the global export name that should be associated with the
mainModuleFile for libraries that has an iife dist build. The fileExtensions array entry defines the
list of file extensions supported. all other files are regarded as assets which can be copied over
during the build process