feat: add more config options to our build localized config file. enable separated dist and lib buil
The "copyAssets" option if set to true instructs the module to copy asset files into the final build
folders. Defaults to false. Asset files are those files whose extension is not among the list of
fileExtensions config entry. The "uglify" option instructs the module to produce minified version of
each module. Defaults to false. minified version has .min appended to their file name. The
uglifyOnly option determines if only the uglified versions should be created. Defaults to false. The
interop option specifies if Rollup js should enable the interop build option. defaults to true.
sourcemap defaults to true, it can be true, 'inline' or false. The distConfig option defines options
that are specific to the dist build. the libConfig option defines options that are specific to the
lib build. Dist builds are builds that can be distributed independently such as browser iife build,
while lib builds are build meant to be used as part of compenents such as node.js cjs builds.