feat: add a getConfig method that returns the final build config for each module's build kind cum ug
The getConfig method returns the config for a module build kind, the build kind can either be a dist
unminified build, a dist minified build, a lib unminified build or a lib minified build. it is the
responsibility of the getExports method to drive these decisions based on the config options
defined.