feat: create the main export/entry file. these is the API facing end that is used to interact with t
it exports an object with a solitary getExports method that drives the process and returns an array
of config options as required by rollup.js. it takes an optional uglifier plugin as first parameter,
(set null if no uglifier plugin is to to be used), an array of other optional plugins to use as a
second parameter, and an optional configPath string that specifies the user defined build config
file to parse. it defaults to .buildrc.json file located in your project root folder. The path
should be relative to your project root folder. example of using this bundler.getExports(uglify(),
[json()], 'myConfigFolder/config.json') . etc